@@ -37,17 +37,30 @@
             "homepage": "http://superdweebie.com"
         }
     ],
+    "repositories": [
+        {
+            "type": "vcs",
+            "url": "https://github.com/webimpress/DoctrineModule"
+        }
+    ],
     "require": {
-        "php": "~5.5 || ~7.0",
-        "doctrine/doctrine-module": "~1.0",
-        "doctrine/mongodb-odm": "~1.0",
-        "zendframework/zend-mvc": "2.*",
-        "zendframework/zend-servicemanager": "2.*",
-        "zendframework/zend-stdlib": "2.*"
+        "php":                               "~5.6 || ~7.0",
+        "doctrine/doctrine-module":          "dev-feature/zf3compat",
+        "doctrine/mongodb-odm":              "^1.1",
+        "zendframework/zend-mvc":            "^2.7.10 || ^3.0.2",
+        "zendframework/zend-servicemanager": "^2.7.6 || ^3.1.1",
+        "zendframework/zend-stdlib":         "^2.7.7 || ^3.0.1"
     },
     "require-dev": {
-        "zendframework/zendframework": "2.*",
-        "phpunit/phpunit": "~4.8 || ~5.0"
+        "phpunit/phpunit":                  "^4.8",
+        "squizlabs/php_codesniffer":        "^2.6.2",
+        "zendframework/zend-console":       "^2.6",
+        "zendframework/zend-i18n":          "^2.7.3",
+        "zendframework/zend-log":           "^2.9",
+        "zendframework/zend-modulemanager": "^2.7.2",
+        "zendframework/zend-serializer":    "^2.8",
+        "zendframework/zend-session":       "^2.7.3",
+        "zendframework/zend-view":          "^2.8.1"
     },
     "autoload": {
         "psr-0": {
