@@ -21,6 +21,7 @@ namespace DoctrineMongoODMModule\Service;
 use DoctrineModule\Service\AbstractFactory;
 use Doctrine\ODM\MongoDB\Configuration;
 use Doctrine\ODM\MongoDB\Types\Type;
+use Interop\Container\ContainerInterface;
 use Zend\ServiceManager\ServiceLocatorInterface;
 
 /**
@@ -33,22 +34,21 @@ use Zend\ServiceManager\ServiceLocatorInterface;
  */
 class ConfigurationFactory extends AbstractFactory
 {
-
     /**
-     * @param \Zend\ServiceManager\ServiceLocatorInterface $serviceLocator
-     * @return \Doctrine\ODM\MongoDB\Configuration
-     * @throws \Exception
+     * {@inheritDoc}
+     *
+     * @return Configuration
      */
-    public function createService(ServiceLocatorInterface $serviceLocator)
+    public function __invoke(ContainerInterface $container, $requestedName, array $options = null)
     {
         /** @var $options \DoctrineMongoODMModule\Options\Configuration */
-        $options = $this->getOptions($serviceLocator, 'configuration');
+        $options = $this->getOptions($container, 'configuration');
 
         $config = new Configuration;
 
         // logger
         if ($options->getLogger()) {
-            $logger = $serviceLocator->get($options->getLogger());
+            $logger = $container->get($options->getLogger());
             $config->setLoggerCallable(array($logger, 'log'));
         }
 
@@ -66,22 +66,22 @@ class ConfigurationFactory extends AbstractFactory
         $config->setDefaultDB($options->getDefaultDb());
 
         // caching
-        $config->setMetadataCacheImpl($serviceLocator->get($options->getMetadataCache()));
+        $config->setMetadataCacheImpl($container->get($options->getMetadataCache()));
 
         // retries
         $config->setRetryConnect($options->getRetryConnect());
         $config->setRetryQuery($options->getRetryQuery());
 
         // Register filters
-        foreach($options->getFilters() as $alias => $class){
+        foreach ($options->getFilters() as $alias => $class) {
             $config->addFilter($alias, $class);
         }
 
         // the driver
-        $config->setMetadataDriverImpl($serviceLocator->get($options->getDriver()));
+        $config->setMetadataDriverImpl($container->get($options->getDriver()));
 
         // metadataFactory, if set
-        if ($factoryName = $options->getClassMetadataFactoryName()){
+        if ($factoryName = $options->getClassMetadataFactoryName()) {
             $config->setClassMetadataFactoryName($factoryName);
         }
 
@@ -95,6 +95,11 @@ class ConfigurationFactory extends AbstractFactory
         }
 
         return $config;
+    }
+
+    public function createService(ServiceLocatorInterface $container)
+    {
+        return $this($container, Configuration::class);
     }
 
     public function getOptionsClass()
