@@ -16,10 +16,11 @@
  * and is licensed under the MIT license. For more information, see
  * <http://www.doctrine-project.org>.
  */
-namespace DoctrineMongoODMModuleTest\Doctrine\Util;
+namespace DoctrineMongoODMModuleTest;
 
-use Zend\ServiceManager\ServiceManager;
+use Zend\Mvc\Application;
 use Zend\Mvc\Service\ServiceManagerConfig;
+use Zend\ServiceManager\ServiceManager;
 
 /**
  * Utility used to retrieve a freshly bootstrapped application's service manager
@@ -31,41 +32,38 @@ use Zend\Mvc\Service\ServiceManagerConfig;
 class ServiceManagerFactory
 {
     /**
-     * @var array
+     * @return array
      */
-    protected static $config;
-
-    /**
-     * @param array $config
-     */
-    public static function setConfig(array $config)
+    public static function getConfiguration()
     {
-        static::$config = $config;
+        $r = new \ReflectionClass(Application::class);
+        $requiredParams = $r->getConstructor()->getNumberOfRequiredParameters();
+
+        $configFile = $requiredParams == 1 ? 'TestConfigurationV3.php' : 'TestConfigurationV2.php';
+
+        return include __DIR__ . '/../' . $configFile;
     }
 
     /**
      * Builds a new service manager
      *
-     * @return \Zend\ServiceManager\ServiceManager
+     * @param  array|null     $configuration
+     * @return ServiceManager
      */
-    public static function getServiceManager()
+    public static function getServiceManager(array $configuration = null)
     {
-        $serviceManager = new ServiceManager(
-            new ServiceManagerConfig(
-                isset(static::$config['service_manager']) ? static::$config['service_manager'] : array()
-            )
+        $configuration        = $configuration ?: static::getConfiguration();
+        $serviceManager       = new ServiceManager();
+        $serviceManagerConfig = new ServiceManagerConfig(
+            isset($configuration['service_manager']) ? $configuration['service_manager'] : array()
         );
-        $serviceManager->setService('ApplicationConfig', static::$config);
-        if (!$serviceManager->has('ServiceListener')) {
-            $serviceManager->setFactory(
-                'ServiceListener',
-                'Zend\Mvc\Service\ServiceListenerFactory'
-            );
-        }
+        $serviceManagerConfig->configureServiceManager($serviceManager);
+        $serviceManager->setService('ApplicationConfig', $configuration);
 
         /** @var $moduleManager \Zend\ModuleManager\ModuleManager */
         $moduleManager = $serviceManager->get('ModuleManager');
         $moduleManager->loadModules();
+
         return $serviceManager;
     }
 }
